feat: Add Dockerfile for Redis service
This commit adds a Dockerfile for the Redis service. It uses the official Redis image as the base image and sets a non-root user for better security. The entrypoint is set to run the Redis server. This change is made to improve the deployment and management of the Redis service.