feat: Add container_image attribute to Rustscan class
This commit adds a new attribute, `container_image`, to the Rustscan class in the `port_scanners.py` file. The `container_image` attribute is set to "rustscan/rustscan" and represents the Docker container image used for running Rustscan scans. This change enhances the functionality of the application by providing a more flexible and customizable option for specifying the container image for Rustscan scans.